requirements.txt
apscheduler==3.9.1.post1

asgiref==3.5.2

azure_storage==0.36.0

boto==2.44.0

boto3==1.26.12

Django==3.2.2

djangorestframework==3.14.0

oauth2client==4.1.3

PyJWT==2.6.0

pytz==2022.6

requests==2.27.1

urllib3         ==          1.26.13

importlib-metadata==4.13.0

#django-clamav       ==          0.4.0

django-cors-headers   ==        3.13.0

django-rest-swagger    ==       2.2.0

djangorestframework     ==      3.14.0

djangorestframework-simplejwt ==5.2.2

wheel               ==          0.38.4

celery            ==            5.2.7

psycopg2            ==          2.9.5